dairy_poc/requirements.txt
# Core data stack
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# Visualisation
plotly>=5.18.0

# ML / anomaly detection
scikit-learn>=1.3.0

# Serialisation
pyarrow>=14.0.0
pyyaml>=6.0.1

# Synthetic data
faker>=22.0.0

# App (scaffolded for later)
streamlit>=1.31.0
